# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ImageProcessor API
  version: 0.1.0
paths:
  /images:
    get:
      summary: List uploaded images and processing statuses
      responses:
        "200":
          description: Image list
  /upload:
    post:
      summary: Upload image and enqueue background processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "202":
          description: Image accepted
        "400":
          description: Invalid upload
  /image/{id}:
    get:
      summary: Get processed image or current processing state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Processed image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "202":
          description: Image is not processed yet
        "404":
          description: Image not found
    delete:
      summary: Delete image and metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /image/{id}/thumbnail:
    get:
      summary: Get processed thumbnail or current processing state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Processed thumbnail
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "202":
          description: Image is not processed yet
        "404":
          description: Image not found
  /status/{id}:
    get:
      summary: Get image processing status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image status
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: API is healthy